Use k8snet.IPFamily in the netlink wrapper methods (#3381)
* Use k8snet.IPFamily in netlink.RouteList wrapper

The underlying lib API takes an int for the family so change to
k8snet.IPFamily in the wrapper for better typing. Most current
call sites were using syscall.AF_INET which technically should have
used the FAMILY_V4 constant defined in the underlying lib (both are
defined as 2).

Signed-off-by: Tom Pantelis <tompantelis@gmail.com>

* Use k8snet.IPFamily in netlink.AddrList wrapper

The underlying lib API takes an int for the family so change to
k8snet.IPFamily in the wrapper for better typing.

Signed-off-by: Tom Pantelis <tompantelis@gmail.com>

* Use k8snet.IPFamily in netlink.RuleList wrapper

The underlying lib API takes an int for the family so change to
k8snet.IPFamily in the wrapper for better typing.

Signed-off-by: Tom Pantelis <tompantelis@gmail.com>

* Use k8snet.IPFamily in netlink.DeleteXfrmRules wrapper

The underlying lib API takes an int for the family so change to
k8snet.IPFamily in the wrapper for better typing.

Signed-off-by: Tom Pantelis <tompantelis@gmail.com>

* Use k8snet.IPFamily in netlink.GetDefaultGatewayInterface

The underlying lib API takes an int for the family so change to
k8snet.IPFamily in the wrapper for better typing.

Signed-off-by: Tom Pantelis <tompantelis@gmail.com>

---------

Signed-off-by: Tom Pantelis <tompantelis@gmail.com>
Co-authored-by: Yossi Boaron <yboaron@redhat.com>