Add unit test hook for net.Dial used for obtaining local IP
Since net.Dial may not work for IPv6 in unit tests and we're mocking
netlink for the fallback method, we should avoid net.Dial altogether
and just mock it to return an error. We keep one test case that uses
the real net.Dial for IPv4 to cover the code path.

Signed-off-by: Tom Pantelis <tompantelis@gmail.com>