Fix ip6tables binary provisioning in containers (#3422)
* Fix ip6tables binary provisioning in containers

The iptables wrapper installer used an alternatives setup which no
longer works correctly on Fedora 41, and ended up skipping ip6tables
(and even removing it, which is why the earlier link is removed).
This updates the alternatives setup to use different names for the
ip6tables alternatives, restoring ip6tables features before and after
the wrapper is called.

Signed-off-by: Stephen Kitt <skitt@redhat.com>

* Enable the IPv6-only variant of E2E

This ensures that IPv6-only is tested in this project; this avoids
IPv6-only issues only showing up in other projects.

Signed-off-by: Stephen Kitt <skitt@redhat.com>

---------

Signed-off-by: Stephen Kitt <skitt@redhat.com>