Ignore link index 0 in GetDefaultGatewayInterface
Previously it returned an error if it found a route with
`LinkIndex == 0` which caused an intermittent unit test failure.
The unit tests first set up a default gateway which adds a route
with link index 99 and, most of the time, this is the first one
returned from RouteList. But the fake netlink implementation
stores the routes in a map keyed by link index so sometimes
RouteList returns the first route with link index 0 since iterating
a map occurs in random order.

So instead of GetDefaultGatewayInterface returning an error if a
route has a 0 LinkIndex, simply ignore it.

Signed-off-by: Tom Pantelis <tompantelis@gmail.com>