Implement intra-cluster routing disablement in the OVN handler
If intra-cluster routing is disabled then elide processing
NonGatewayRoutes, ie don't instantiate the NonGatewayRouteController
and NonGatewayRouteHandler.

In addition, only program the host networking dataplane if the local
node is the active gateway node. So on a passive gateway node, don't add
host networking rules/routes for remote Endpoints. On transition to
the the active gateway, then add the rules/routes for existing
Endpoints. On transition to passive, remove the rules/routes for existing
Endpoints.

Signed-off-by: Tom Pantelis <tompantelis@gmail.com>