Improve close detection and check for truncated messages
...in natDiscovery.listenerLoop

The Cursor AI tool identified a couple potential issues:

The code assumes that length == 0 from ReadFromUDP only means the connection
is closed. However, UDP can legitimately receive zero-length packets. To
be safe, check the returned error for net.ErrClosed.

A fixed buffer size of 2048 is used but a UDP message larger than that
will be silently truncated which would cause a parsing error. Add a warning
re: truncation if the returned length is 2048.

Signed-off-by: Tom Pantelis <tompantelis@gmail.com>