Remove unused netlink.Adapter methods
The Cursor AI tool identified a logic error in AddrAddIfNotPresent but,
as it turns out, this method is actually not used anywhere so we can
remove it. Also AddDestinationRoutes and DeleteDestinationRoutes aren't
used either.

These were added by https://github.com/submariner-io/submariner/pull/1741
to support a new cable driver implementation that was abandoned.

Signed-off-by: Tom Pantelis <tompantelis@gmail.com>