Handle equal CreationTimestamps for Endpoint removal events
We check that the timestamp of the prior Endpoint for a cluster on
removal is after/later than the removed Endpoint but, due to seconds
granularity of CreationTimestamps in K8s, they could actually be equal.
This was observed in a production environment, which caused handlers to
be notified of a non-stale remove event.

Instead of checking CreationTimestamps, we can check the UIDs - only
notify of removal if they match, otherwise consider it stale,

See related issue https://issues.redhat.com/browse/ACM-25262.

Signed-off-by: Tom Pantelis <tompantelis@gmail.com>